clean up of snow routines, ran uncrustify and moved frequently used formulas to functions in snow_utility.c

+++ src/snow_intercept.c
@@ -467,7 +467,7 @@ int snow_intercept(double  Dt,
   }
 
   if ( *IntSnow <= 0 )
-    RainThroughFall = veg_var->throughfall / MMPERMETER;
+    RainThroughFall = veg_var->throughfall / (double)MMPERMETER;
 
   RefreezeEnergy *= Dt;
